synchronize

+++ src/dfx/src/actors/pocketic_proxy.rs
@@ -247,26 +247,23 @@ fn pocketic_proxy_start_thread(
                 .expect("Could not write to pocketic-proxy-pid file.");
             let port =
                 PocketIcProxy::wait_for_ready(&pocketic_proxy_port_path, &ready_file).unwrap();
-            let _instance = match block_on_initialize_gateway(
+            if let Err(e) = block_on_initialize_gateway(
                 format!("http://localhost:{port}").parse().unwrap(),
                 replica_url.clone(),
                 domains.clone(),
                 address,
                 logger.clone(),
             ) {
-                Err(e) => {
-                    error!(logger, "Failed to initialize HTTP gateway: {:#}", e);
-                    let _ = child.kill();
-                    let _ = child.wait();
-                    if receiver.try_recv().is_ok() {
-                        debug!(logger, "Got signal to stop.");
-                        break;
-                    } else {
-                        continue;
-                    }
+                error!(logger, "Failed to initialize HTTP gateway: {:#}", e);
+                let _ = child.kill();
+                let _ = child.wait();
+                if receiver.try_recv().is_ok() {
+                    debug!(logger, "Got signal to stop.");
+                    break;
+                } else {
+                    continue;
                 }
-                Ok(i) => i,
-            };
+            }
             info!(logger, "Replica API running on {address}");
 
             // This waits for the child to stop, or the receiver to receive a message.
@@ -310,7 +307,7 @@ fn block_on_initialize_gateway(
     domains: Vec<String>,
     addr: SocketAddr,
     logger: Logger,
-) -> DfxResult<usize> {
+) -> DfxResult {
     Builder::new_current_thread()
         .enable_all()
         .build()
@@ -331,12 +328,13 @@ async fn initialize_gateway(
     domains: Vec<String>,
     addr: SocketAddr,
     logger: Logger,
-) -> DfxResult<usize> {
+) -> DfxResult {
     use pocket_ic::common::rest::{
         CreateHttpGatewayResponse, HttpGatewayBackend, HttpGatewayConfig,
     };
     use reqwest::Client;
     let init_client = Client::new();
+    debug!(logger, "Configuring PocketIC gateway");
     let resp = init_client
         .post(pocketic_url.join("http_gateway").unwrap())
         .json(&HttpGatewayConfig {
@@ -350,12 +348,11 @@ async fn initialize_gateway(
         .await?
         .error_for_status()?;
     let resp = resp.json::<CreateHttpGatewayResponse>().await?;
-    let instance = match resp {
-        CreateHttpGatewayResponse::Error { message } => bail!("Gateway init error: {message}"),
-        CreateHttpGatewayResponse::Created(info) => info.instance_id,
-    };
+    if let CreateHttpGatewayResponse::Error { message } = resp {
+        bail!("Gateway init error: {message}")
+    }
     info!(logger, "Initialized HTTP gateway.");
-    Ok(instance)
+    Ok(())
 }
 
 #[cfg(not(unix))]
